Decouple GamePage into a Modular Architecture
