+++ src/me3_manager/ui/main_window.py
@@ -377,7 +377,7 @@ class ModEngine3Manager(QMainWindow):
         self.game_container.set_game_order(game_order)
 
         # 5. Rebuild the game pages in the content area
-        from me3_manager.ui.game_page import GamePage
+        from me3_manager.ui.game_page_components import GamePage
 
         # Iterate through the ordered list to add pages sequentially
         for game_name in game_order:
